esling and build fixes READY TO BUILD
