+++ package.json
@@ -65,6 +65,7 @@
     "@types/react": "^18.2.15",
     "@types/react-dom": "^18.2.7",
     "@types/react-modal": "^3.16.3",
+    "@types/react-sparklines": "^1.7.5",
     "@types/sockjs-client": "^1.5.4",
     "@typescript-eslint/eslint-plugin": "^6.0.0",
     "@typescript-eslint/parser": "^6.0.0",
