+++ src/pages/profilePage/ui/components/profileWallet/profileWallet.tsx
@@ -1,15 +1,15 @@
 import { FC } from "react";
-import { PagedRequest } from "entities/request/intex";
-import profileWallet from 'shared/assets/icons/profileWallet.svg'
-import { useGetRequest } from "shared/lib/hooks/useGetRequest";
-import { Button } from "shared/ui/button";
+import { Link } from "react-router-dom";
 
-import { fetchWalletBalance, fetchWalletTransactions } from "../../model/sevices/walletServices";
+// import { PagedRequest } from "entities/request/intex";
+// import profileWallet from 'shared/assets/icons/profileWallet.svg'
+// import { useGetRequest } from "shared/lib/hooks/useGetRequest";
+// import { Button } from "shared/ui/button";
+// import { fetchWalletBalance, fetchWalletTransactions } from "../../model/sevices/walletServices";
+import WalletBalance from "../walletBalance/walletBalance";
 import WalletTrasaction from "../walletTransaction/walletTransaction";
 
 import './profileWallet.scss'
-import { Link } from "react-router-dom";
-import WalletBalance from "../walletBalance/walletBalance";
 
 interface Props {
     userId: string
@@ -23,34 +23,26 @@ export interface Transaction {
     type: string
 }
 
-export interface WalletBalance {
-    balance: number,
-    bonusBalance: number,
-    updatedAt: string,
-    userId: string,
-    walletId: string
-}
+// export interface WalletBalance {
+//     balance: number,
+//     bonusBalance: number,
+//     updatedAt: string,
+//     userId: string,
+//     walletId: string
+// }
 
-// тестовые данные
-const transaction = {
-    amount: 1000,
-    createdAr: [10, 10, 10, 10, 10],
-    id: '1',
-    direction: 'dadada',
-    type: 'dddd'
-}
 
-const ProfileWallet: FC <Props>= ({userId}) =>{
+const ProfileWallet: FC <Props>= () =>{
 
     // запрос на получение данных кошелька
-    const {data: wallet, isLoaded: walletLoaded} = useGetRequest<WalletBalance | string>({fetchFunc: () => fetchWalletBalance(userId), enabled: true, key: []})
-    const {data: transactions, isLoaded: transactionsLoaded} = useGetRequest<PagedRequest<Transaction> | string>({fetchFunc: () => fetchWalletTransactions(userId), enabled: true, key: []})
+    // const {data: wallet, isLoaded: walletLoaded} = useGetRequest<WalletBalance | string>({fetchFunc: () => fetchWalletBalance(userId), enabled: true, key: []})
+    // const {data: transactions, isLoaded: transactionsLoaded} = useGetRequest<PagedRequest<Transaction> | string>({fetchFunc: () => fetchWalletTransactions(userId), enabled: true, key: []})
 
     // if(typeof(wallet) === "string"){
     //     return <div>ERROR WALLET</div>
     // }
 
-    const transactionsList = (transactions && typeof transactions !== 'string') ? transactions.content : [];
+    // const transactionsList = (transactions && typeof transactions !== 'string') ? transactions.content : [];
     
     return(
         <div className="profileWallet">
