+++ src/shared/ui/tag/tag.tsx
@@ -37,10 +37,6 @@ export default function Tag(props: TagProps) {
             
     }
 
-    if(statusTranslation(type) === 'Inactive'){
-        console.log(type)
-    }
-
     return (
         <Text
             Tag='span'
